Most things now go through the Array() structure (transformation matrices etc) so they can be rounded in the same place
