+++ io_scene_godot/structures.py
@@ -4,7 +4,7 @@ This file contains classes to help dealing with the actual writing to the file
 """
 import os
 import collections
-from .encoders import CONVERSIONS
+import mathutils
 
 
 class ESCNFile:
@@ -122,13 +122,7 @@ class FileEntry(collections.OrderedDict):
         out_str = ''
         for var in self:
             val = self[var]
-
-            converter = CONVERSIONS.get(type(val))
-            if converter is not None:
-                val = converter(val)
-
-            if hasattr(val, "to_string"):
-                val = val.to_string()
+            val = to_string(val)
 
             out_str += '\n{} = {}'.format(var, val)
         return out_str
@@ -269,7 +263,7 @@ class Array(list):
         """Convert the array to serialized form"""
         return "{}{}{}".format(
             self.prefix,
-            self.seperator.join([str(v) for v in self]),
+            self.seperator.join([to_string(v) for v in self]),
             self.suffix
         )
 
@@ -287,3 +281,91 @@ class NodePath:
             self.ref_path,
             self.attr_name
         )
+
+
+# ------------------ Implicit Conversions of Blender Types --------------------
+def mat4_to_string(mtx):
+    """Converts a matrix to a "Transform" string that can be parsed by Godot"""
+    def fix_matrix(mtx):
+        """ Shuffles a matrix to change from y-up to z-up"""
+        # Todo: can this be replaced my a matrix multiplcation?
+        trans = mathutils.Matrix(mtx)
+        up_axis = 2
+
+        for i in range(3):
+            trans[1][i], trans[up_axis][i] = trans[up_axis][i], trans[1][i]
+        for i in range(3):
+            trans[i][1], trans[i][up_axis] = trans[i][up_axis], trans[i][1]
+
+        trans[1][3], trans[up_axis][3] = trans[up_axis][3], trans[1][3]
+
+        trans[up_axis][0] = -trans[up_axis][0]
+        trans[up_axis][1] = -trans[up_axis][1]
+        trans[0][up_axis] = -trans[0][up_axis]
+        trans[1][up_axis] = -trans[1][up_axis]
+        trans[up_axis][3] = -trans[up_axis][3]
+
+        return trans
+
+    mtx = fix_matrix(mtx)
+    array = Array('Transform(')
+    for row in range(3):
+        for col in range(3):
+            array.append(mtx[row][col])
+
+    # Export the basis
+    for axis in range(3):
+        array.append(mtx[axis][3])
+
+    return array.to_string()
+
+
+def color_to_string(rgba):
+    """Converts an RGB colors in range 0-1 into a fomat Godot can read. Accepts
+    iterables of 3 or 4 in length, but is designed for mathutils.Color"""
+    alpha = 1.0 if len(rgba) < 4 else rgba[3]
+    col = list(rgba[0:3]) + [alpha]
+    return Array('Color(', values=[col]).to_string()
+
+
+def vector_to_string(vec):
+    """Encode a mathutils.vector. actually, it accepts iterable of any length,
+    but 2, 3 are best...."""
+    return Array('Vector{}('.format(len(vec)), values=[vec]).to_string()
+
+
+def float_to_string(num):
+    """Intelligently rounds float numbers"""
+    if abs(num) < 1e-15:
+        # This should make floating point errors round sanely. It does mean
+        # that if you have objects with large scaling factors and tiny meshes,
+        # then the object may "collapse" to zero.
+        # There are still some e-8's that appear in the file, but I think
+        # people would notice it collapsing.
+        return '0.0'
+    return '{:.6}'.format(num)
+
+
+def to_string(val):
+    """Attempts to convert any object into a string using the conversions
+    table, explicit conversion, or falling back to the str() method"""
+    if hasattr(val, "to_string"):
+        val = val.to_string()
+    else:
+        converter = CONVERSIONS.get(type(val))
+        if converter is not None:
+            val = converter(val)
+        else:
+            val = str(val)
+
+    return val
+
+
+# Finds the correct conversion function for a datatype
+CONVERSIONS = {
+    float: float_to_string,
+    bool: lambda x: 'true' if x else 'false',
+    mathutils.Matrix: mat4_to_string,
+    mathutils.Color: color_to_string,
+    mathutils.Vector: vector_to_string
+}
