+++ tests/reference-exports/just_armature.escn
@@ -6,34 +6,34 @@
 
 [node name="Armature" type="Skeleton" parent="."]
 
-transform = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 0.0, 0.0, -0.0 )
+transform = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0)
 bones/0/name = "Bone"
 bones/0/parent = -1
-bones/0/rest = Transform(  1.0, 0.0, -0.0, 0.0, 0.0, -1.0, -0.0, 1.0, 0.0, 0.0, 0.0, -0.0 )
-bones/0/pose = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 0.0, 0.0, -0.0 )
+bones/0/rest = Transform(1.0, 0.0, 0.0, 0.0, 0.0, -1.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0)
+bones/0/pose = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0)
 bones/0/enabled = true
 bones/1/name = "Bone.001"
 bones/1/parent = 0
-bones/1/rest = Transform(  -1.0, 8.742277657347586e-08, -0.0, 0.0, 0.0, -1.0, -8.742277657347586e-08, -1.0, 0.0, 0.0, 0.0, -1.0 )
-bones/1/pose = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 0.0, 0.0, -0.0 )
+bones/1/rest = Transform(-1.0, 8.74228e-08, 0.0, 0.0, 0.0, -1.0, -8.74228e-08, -1.0, 0.0, 0.0, 0.0, -1.0)
+bones/1/pose = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0)
 bones/1/enabled = true
 bones/2/name = "Bone.002"
 bones/2/parent = 1
-bones/2/rest = Transform(  -1.0, -6.357302595461078e-08, -8.742277657347586e-08, 8.742277657347586e-08, -2.0850114102149812e-15, -1.0, 6.357302595461078e-08, -1.0, 0.0, 0.0, 0.0, -1.0 )
-bones/2/pose = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 0.0, 0.0, -0.0 )
+bones/2/rest = Transform(-1.0, -6.3573e-08, -8.74228e-08, 8.74228e-08, -2.08501e-15, -1.0, 6.3573e-08, -1.0, 0.0, 0.0, 0.0, -1.0)
+bones/2/pose = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0)
 bones/2/enabled = true
 bones/3/name = "Bone.003"
 bones/3/parent = 0
-bones/3/rest = Transform(  0.0, 0.0, -0.9999999403953552, 1.0, -4.371138828673793e-08, -0.0, -4.371138828673793e-08, -1.0, 0.0, 0.0, 0.0, -1.0 )
-bones/3/pose = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 0.0, 0.0, -0.0 )
+bones/3/rest = Transform(0.0, 0.0, -1.0, 1.0, -4.37114e-08, 0.0, -4.37114e-08, -1.0, 0.0, 0.0, 0.0, -1.0)
+bones/3/pose = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0)
 bones/3/enabled = true
 
 [node name="Lamp" type="OmniLight" parent="."]
 
-omni_range = 29.999982833862305
+omni_range = 30.0
 shadow_enabled = true
-light_color = Color( 1.0, 1.0, 1.0, 1.0 )
-transform = Transform(  -0.29086464643478394, -0.7711008191108704, 0.5663931369781494, -0.05518905818462372, 0.6045246720314026, 0.7946722507476807, -0.9551711678504944, 0.199883371591568, -0.21839118003845215, 4.076245307922363, 5.903861999511719, -1.0054539442062378 )
+light_color = Color(1.0, 1.0, 1.0, 1.0)
+transform = Transform(-0.290865, -0.771101, 0.566393, -0.0551891, 0.604525, 0.794672, -0.955171, 0.199883, -0.218391, 4.07625, 5.90386, -1.00545)
 light_negative = false
 light_specular = 1.0
 light_energy = 1.0
@@ -41,7 +41,7 @@ light_energy = 1.0
 [node name="Camera" type="Camera" parent="."]
 
 far = 100.0
-near = 0.10000000149011612
+near = 0.1
 projection = 0
-fov = 49.13434207760448
-transform = Transform(  0.6859206557273865, -0.32401350140571594, 0.6515582203865051, 0.0, 0.8953956365585327, 0.44527143239974976, -0.7276763319969177, -0.3054208755493164, 0.6141703724861145, 7.481131553649902, 5.34366512298584, 6.5076398849487305 )
+fov = 49.1343
+transform = Transform(0.685921, -0.324014, 0.651558, 0.0, 0.895396, 0.445271, -0.727676, -0.305421, 0.61417, 7.48113, 5.34367, 6.50764)
