+++ tests/reference-exports/just_cameras.escn
@@ -9,37 +9,37 @@
 far = 10.0
 near = 5.0
 projection = 0
-fov = 90.00000250447816
-transform = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 5.0, 0.0, -0.0 )
+fov = 90.0
+transform = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 5.0, 0.0, 0.0)
 
 [node name="Camera003" type="Camera" parent="."]
 
 far = 100.0
-near = 0.10000000149011612
+near = 0.1
 projection = 1
 size = 6.0
-transform = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 1.0, 0.0, -0.0 )
+transform = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0)
 
 [node name="Camera002" type="Camera" parent="."]
 
 far = 100.0
-near = 0.10000000149011612
+near = 0.1
 projection = 1
 size = 2.0
-transform = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, -2.0, 0.0, -0.0 )
+transform = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, -2.0, 0.0, 0.0)
 
 [node name="Camera001" type="Camera" parent="."]
 
 far = 100.0
-near = 0.10000000149011612
+near = 0.1
 projection = 0
-fov = 90.00000250447816
-transform = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, -5.0, 0.0, -0.0 )
+fov = 90.0
+transform = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, -5.0, 0.0, 0.0)
 
 [node name="Camera" type="Camera" parent="."]
 
 far = 100.0
-near = 0.10000000149011612
+near = 0.1
 projection = 0
-fov = 19.99999941818584
-transform = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, -7.0, 0.0, -0.0 )
+fov = 20.0
+transform = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, -7.0, 0.0, 0.0)
