+++ tests/reference-exports/just_point_lights.escn
@@ -6,7 +6,7 @@ surfaces/0 = {
 	"primitive":4,
 	"arrays":[
 		Vector3Array(-1.0, 1.0, 1.0, -1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, 1.0, -1.0, 1.0, -1.0, -1.0, -1.0, -1.0, -1.0, 1.0, 1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -1.0, -1.0, 1.0, 1.0, 1.0, -1.0, -1.0, 1.0, 1.0, -1.0, 1.0, 1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, 1.0, 1.0, 1.0, 1.0, 1.0, -1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -1.0, 1.0, 1.0, 1.0, -1.0, 1.0, 1.0, 1.0, -1.0, 1.0, -1.0, 1.0, 1.0),
-		Vector3Array(-1.0, 0.0, -0.0, -1.0, 0.0, -0.0, -1.0, 0.0, -0.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 1.0, 0.0, -0.0, 1.0, 0.0, -0.0, 1.0, 0.0, -0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, -1.0, -0.0, 0.0, -1.0, -0.0, 0.0, -1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0, -1.0, 0.0, -0.0, 0.0, 0.0, -1.0, 1.0, 0.0, -0.0, 0.0, 0.0, 1.0, 0.0, -1.0, -0.0, 0.0, 1.0, -0.0),
+		Vector3Array(-1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, -1.0, 0.0, 0.0, 0.0, 0.0, -1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, -1.0, 0.0, 0.0, 1.0, 0.0),
 		null, ; No Tangents,
 		null, ; no Vertex Colors,
 		null, ; No UV1,
@@ -25,7 +25,7 @@ surfaces/0 = {
 	"primitive":4,
 	"arrays":[
 		Vector3Array(-1.0, 1.0, 1.0, -1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, 1.0, -1.0, 1.0, -1.0, -1.0, -1.0, -1.0, -1.0, 1.0, 1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -1.0, -1.0, 1.0, 1.0, 1.0, -1.0, -1.0, 1.0, 1.0, -1.0, 1.0, 1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, 1.0, 1.0, 1.0, 1.0, 1.0, -1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -1.0, 1.0, 1.0, 1.0, -1.0, 1.0, 1.0, 1.0, -1.0, 1.0, -1.0, 1.0, 1.0),
-		Vector3Array(-1.0, 0.0, -0.0, -1.0, 0.0, -0.0, -1.0, 0.0, -0.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 1.0, 0.0, -0.0, 1.0, 0.0, -0.0, 1.0, 0.0, -0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, -1.0, -0.0, 0.0, -1.0, -0.0, 0.0, -1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0, -1.0, 0.0, -0.0, 0.0, 0.0, -1.0, 1.0, 0.0, -0.0, 0.0, 0.0, 1.0, 0.0, -1.0, -0.0, 0.0, 1.0, -0.0),
+		Vector3Array(-1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, -1.0, 0.0, 0.0, 0.0, 0.0, -1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, -1.0, 0.0, 0.0, 1.0, 0.0),
 		null, ; No Tangents,
 		null, ; no Vertex Colors,
 		null, ; No UV1,
@@ -44,7 +44,7 @@ surfaces/0 = {
 	"primitive":4,
 	"arrays":[
 		Vector3Array(1.0, 0.0, 1.0, -1.0, 0.0, -1.0, -1.0, 0.0, 1.0, 1.0, 0.0, -1.0),
-		Vector3Array(0.0, 1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0),
+		Vector3Array(0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0),
 		null, ; No Tangents,
 		null, ; no Vertex Colors,
 		null, ; No UV1,
@@ -63,8 +63,8 @@ surfaces/0 = {
 
 omni_range = 5.0
 shadow_enabled = true
-light_color = Color( 0.674415111541748, 0.0003378598776180297, 1.0, 1.0 )
-transform = Transform(  1.0, 0.0, -0.0, 0.0, -4.371138828673793e-08, 1.0, -0.0, -1.0, -4.371138828673793e-08, 0.0, 0.3999999761581421, 1.0 )
+light_color = Color(0.674415, 0.00033786, 1.0, 1.0)
+transform = Transform(1.0, 0.0, 0.0, 0.0, -4.37114e-08, 1.0, 0.0, -1.0, -4.37114e-08, 0.0, 0.4, 1.0)
 light_negative = false
 light_specular = 1.0
 light_energy = 1.0
@@ -73,20 +73,20 @@ light_energy = 1.0
 
 mesh = SubResource(1)
 visible = true
-transform = Transform(  0.30000001192092896, 0.0, -0.0, 0.0, 0.6000000238418579, -0.0, -0.0, -0.0, 0.30000001192092896, 0.0, 0.0, 0.35525739192962646 )
+transform = Transform(0.3, 0.0, 0.0, 0.0, 0.6, 0.0, 0.0, 0.0, 0.3, 0.0, 0.0, 0.355257)
 
 [node name="Cube" type="MeshInstance" parent="."]
 
 mesh = SubResource(2)
 visible = true
-transform = Transform(  0.30000001192092896, 0.0, -0.0, 0.0, 0.6000000238418579, -0.0, -0.0, -0.0, 0.30000001192092896, -2.0, 0.0, 0.35525739192962646 )
+transform = Transform(0.3, 0.0, 0.0, 0.0, 0.6, 0.0, 0.0, 0.0, 0.3, -2.0, 0.0, 0.355257)
 
 [node name="Lamp002" type="OmniLight" parent="."]
 
 omni_range = 5.0
 shadow_enabled = false
-light_color = Color( 0.0, 1.0, 0.0024246724788099527, 1.0 )
-transform = Transform(  1.0, 0.0, -0.0, 0.0, -4.371138828673793e-08, 1.0, -0.0, -1.0, -4.371138828673793e-08, -2.0, 0.3999999761581421, 1.0 )
+light_color = Color(0.0, 1.0, 0.00242467, 1.0)
+transform = Transform(1.0, 0.0, 0.0, 0.0, -4.37114e-08, 1.0, 0.0, -1.0, -4.37114e-08, -2.0, 0.4, 1.0)
 light_negative = false
 light_specular = 1.0
 light_energy = 1.0
@@ -95,8 +95,8 @@ light_energy = 1.0
 
 omni_range = 2.0
 shadow_enabled = false
-light_color = Color( 1.0, 0.0, 0.0, 1.0 )
-transform = Transform(  1.0, 0.0, -0.0, 0.0, -4.371138828673793e-08, 1.0, -0.0, -1.0, -4.371138828673793e-08, -6.0, 1.0, 3.0 )
+light_color = Color(1.0, 0.0, 0.0, 1.0)
+transform = Transform(1.0, 0.0, 0.0, 0.0, -4.37114e-08, 1.0, 0.0, -1.0, -4.37114e-08, -6.0, 1.0, 3.0)
 light_negative = false
 light_specular = 1.0
 light_energy = 1.0
@@ -105,14 +105,14 @@ light_energy = 1.0
 
 mesh = SubResource(3)
 visible = true
-transform = Transform(  10.0, 0.0, -0.0, 0.0, 10.0, -0.0, -0.0, -0.0, 10.0, 0.0, 0.0, -0.0 )
+transform = Transform(10.0, 0.0, 0.0, 0.0, 10.0, 0.0, 0.0, 0.0, 10.0, 0.0, 0.0, 0.0)
 
 [node name="Lamp" type="OmniLight" parent="."]
 
 omni_range = 2.0
 shadow_enabled = false
-light_color = Color( 1.0, 1.0, 1.0, 1.0 )
-transform = Transform(  1.0, 0.0, -0.0, 0.0, -4.371138828673793e-08, 1.0, -0.0, -1.0, -4.371138828673793e-08, -6.0, 1.0, -0.0 )
+light_color = Color(1.0, 1.0, 1.0, 1.0)
+transform = Transform(1.0, 0.0, 0.0, 0.0, -4.37114e-08, 1.0, 0.0, -1.0, -4.37114e-08, -6.0, 1.0, 0.0)
 light_negative = false
 light_specular = 1.0
 light_energy = 1.0
@@ -120,7 +120,7 @@ light_energy = 1.0
 [node name="Camera" type="Camera" parent="."]
 
 far = 100.0
-near = 0.10000000149011612
+near = 0.1
 projection = 0
-fov = 49.13434207760448
-transform = Transform(  0.6859206557273865, -0.32401350140571594, 0.6515582203865051, 0.0, 0.8953956365585327, 0.44527143239974976, -0.7276763319969177, -0.3054208755493164, 0.6141703724861145, 7.481131553649902, 5.34366512298584, 6.5076398849487305 )
+fov = 49.1343
+transform = Transform(0.685921, -0.324014, 0.651558, 0.0, 0.895396, 0.445271, -0.727676, -0.305421, 0.61417, 7.48113, 5.34367, 6.50764)
