+++ tests/reference-exports/simple_materials.escn
@@ -6,7 +6,7 @@ flags_unshaded = false
 flags_vertex_lighting = false
 flags_transparent = false
 vertex_color_use_as_albedo = false
-albedo_color = Color( 0.013005342334508896, 0.8000000715255737, 0.0, 1.0 )
+albedo_color = Color(0.0130053, 0.8, 0.0, 1.0)
 subsurf_scatter_enabled = false
 
 [sub_resource id=2 type="SpatialMaterial"]
@@ -15,7 +15,7 @@ flags_unshaded = false
 flags_vertex_lighting = false
 flags_transparent = false
 vertex_color_use_as_albedo = false
-albedo_color = Color( 0.800000011920929, 0.800000011920929, 0.800000011920929, 1.0 )
+albedo_color = Color(0.8, 0.8, 0.8, 1.0)
 subsurf_scatter_enabled = false
 
 [sub_resource id=3 type="ArrayMesh"]
@@ -24,8 +24,8 @@ surfaces/0 = {
 	"material":SubResource(1),
 	"primitive":4,
 	"arrays":[
-		Vector3Array(1.0, -1.0, 1.0, -0.9999996423721313, -1.0, -1.0000003576278687, 1.0, -1.0, -0.9999999403953552, -0.9999999403953552, 1.0, -1.0, 0.9999993443489075, 1.0, 1.0000005960464478, 1.0000004768371582, 1.0, -0.999999463558197, 0.9999993443489075, 1.0, 1.0000005960464478, -1.0000001192092896, -1.0, 0.9999998211860657, 1.0, -1.0, 1.0, -1.0000001192092896, -1.0, 0.9999998211860657, -0.9999999403953552, 1.0, -1.0, -0.9999996423721313, -1.0, -1.0000003576278687, 1.0, -1.0, -0.9999999403953552, -0.9999999403953552, 1.0, -1.0, 1.0000004768371582, 1.0, -0.999999463558197, 1.0, -1.0, 1.0, -1.0000001192092896, -1.0, 0.9999998211860657, -0.9999996423721313, -1.0, -1.0000003576278687, -0.9999999403953552, 1.0, -1.0, -1.0000003576278687, 1.0, 0.9999996423721313, 0.9999993443489075, 1.0, 1.0000005960464478, 0.9999993443489075, 1.0, 1.0000005960464478, -1.0000003576278687, 1.0, 0.9999996423721313, -1.0000001192092896, -1.0, 0.9999998211860657, -1.0000001192092896, -1.0, 0.9999998211860657, -1.0000003576278687, 1.0, 0.9999996423721313, -0.9999999403953552, 1.0, -1.0, 1.0, -1.0, -0.9999999403953552, -0.9999996423721313, -1.0, -1.0000003576278687, -0.9999999403953552, 1.0, -1.0),
-		Vector3Array(2.980232949312267e-08, -1.0, -0.0, 2.980232949312267e-08, -1.0, -0.0, 2.980232949312267e-08, -1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0, 0.0, 1.0, -0.0, -8.940696716308594e-08, -4.76837158203125e-07, 1.0, -8.940696716308594e-08, -4.76837158203125e-07, 1.0, -8.940696716308594e-08, -4.76837158203125e-07, 1.0, -1.0, -1.4901156930591242e-07, -2.3841855067985307e-07, -1.0, -1.4901156930591242e-07, -2.3841855067985307e-07, -1.0, -1.4901156930591242e-07, -2.3841855067985307e-07, 2.6822084464583895e-07, 2.3841852225814364e-07, -1.0, 2.6822084464583895e-07, 2.3841852225814364e-07, -1.0, 2.6822084464583895e-07, 2.3841852225814364e-07, -1.0, 0.0, -1.0, -0.0, 0.0, -1.0, -0.0, 0.0, -1.0, -0.0, 5.96046660916727e-08, 1.0, -0.0, 5.96046660916727e-08, 1.0, -0.0, 5.96046660916727e-08, 1.0, -0.0, -4.768372150465439e-07, 1.1920927533992653e-07, 1.0, -4.768372150465439e-07, 1.1920927533992653e-07, 1.0, -4.768372150465439e-07, 1.1920927533992653e-07, 1.0, -1.0, -1.1920931797249068e-07, -2.3841863594498136e-07, -1.0, -1.1920931797249068e-07, -2.3841863594498136e-07, -1.0, -1.1920931797249068e-07, -2.3841863594498136e-07, 2.0861631355728605e-07, 8.940701690107744e-08, -1.0, 2.0861631355728605e-07, 8.940701690107744e-08, -1.0, 2.0861631355728605e-07, 8.940701690107744e-08, -1.0),
+		Vector3Array(1.0, -1.0, 1.0, -1.0, -1.0, -1.0, 1.0, -1.0, -1.0, -1.0, 1.0, -1.0, 0.999999, 1.0, 1.0, 1.0, 1.0, -0.999999, 0.999999, 1.0, 1.0, -1.0, -1.0, 1.0, 1.0, -1.0, 1.0, -1.0, -1.0, 1.0, -1.0, 1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, -1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -0.999999, 1.0, -1.0, 1.0, -1.0, -1.0, 1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0, -1.0, 1.0, 1.0, 0.999999, 1.0, 1.0, 0.999999, 1.0, 1.0, -1.0, 1.0, 1.0, -1.0, -1.0, 1.0, -1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -1.0, 1.0, -1.0, 1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, 1.0, -1.0),
+		Vector3Array(2.98023e-08, -1.0, 0.0, 2.98023e-08, -1.0, 0.0, 2.98023e-08, -1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, -8.9407e-08, -4.76837e-07, 1.0, -8.9407e-08, -4.76837e-07, 1.0, -8.9407e-08, -4.76837e-07, 1.0, -1.0, -1.49012e-07, -2.38419e-07, -1.0, -1.49012e-07, -2.38419e-07, -1.0, -1.49012e-07, -2.38419e-07, 2.68221e-07, 2.38419e-07, -1.0, 2.68221e-07, 2.38419e-07, -1.0, 2.68221e-07, 2.38419e-07, -1.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 0.0, -1.0, 0.0, 5.96047e-08, 1.0, 0.0, 5.96047e-08, 1.0, 0.0, 5.96047e-08, 1.0, 0.0, -4.76837e-07, 1.19209e-07, 1.0, -4.76837e-07, 1.19209e-07, 1.0, -4.76837e-07, 1.19209e-07, 1.0, -1.0, -1.19209e-07, -2.38419e-07, -1.0, -1.19209e-07, -2.38419e-07, -1.0, -1.19209e-07, -2.38419e-07, 2.08616e-07, 8.9407e-08, -1.0, 2.08616e-07, 8.9407e-08, -1.0, 2.08616e-07, 8.9407e-08, -1.0),
 		null, ; No Tangents,
 		null, ; no Vertex Colors,
 		null, ; No UV1,
@@ -41,8 +41,8 @@ surfaces/1 = {
 	"material":SubResource(2),
 	"primitive":4,
 	"arrays":[
-		Vector3Array(1.0000004768371582, 1.0, -0.999999463558197, 1.0, -1.0, 1.0, 1.0, -1.0, -0.9999999403953552, 1.0000004768371582, 1.0, -0.999999463558197, 0.9999993443489075, 1.0, 1.0000005960464478, 1.0, -1.0, 1.0),
-		Vector3Array(1.0, -2.384185791015625e-07, -0.0, 1.0, -2.384185791015625e-07, -0.0, 1.0, -2.384185791015625e-07, -0.0, 1.0, 3.2782537573439186e-07, 5.960464477539062e-07, 1.0, 3.2782537573439186e-07, 5.960464477539062e-07, 1.0, 3.2782537573439186e-07, 5.960464477539062e-07),
+		Vector3Array(1.0, 1.0, -0.999999, 1.0, -1.0, 1.0, 1.0, -1.0, -1.0, 1.0, 1.0, -0.999999, 0.999999, 1.0, 1.0, 1.0, -1.0, 1.0),
+		Vector3Array(1.0, -2.38419e-07, 0.0, 1.0, -2.38419e-07, 0.0, 1.0, -2.38419e-07, 0.0, 1.0, 3.27825e-07, 5.96046e-07, 1.0, 3.27825e-07, 5.96046e-07, 1.0, 3.27825e-07, 5.96046e-07),
 		null, ; No Tangents,
 		null, ; no Vertex Colors,
 		null, ; No UV1,
@@ -61,14 +61,14 @@ surfaces/1 = {
 
 mesh = SubResource(3)
 visible = true
-transform = Transform(  1.0, 0.0, -0.0, 0.0, 1.0, -0.0, -0.0, -0.0, 1.0, 0.0, 0.0, -0.0 )
+transform = Transform(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0)
 
 [node name="Lamp" type="OmniLight" parent="."]
 
-omni_range = 29.999982833862305
+omni_range = 30.0
 shadow_enabled = true
-light_color = Color( 1.0, 1.0, 1.0, 1.0 )
-transform = Transform(  -0.29086464643478394, -0.7711008191108704, 0.5663931369781494, -0.05518905818462372, 0.6045246720314026, 0.7946722507476807, -0.9551711678504944, 0.199883371591568, -0.21839118003845215, 4.076245307922363, 5.903861999511719, -1.0054539442062378 )
+light_color = Color(1.0, 1.0, 1.0, 1.0)
+transform = Transform(-0.290865, -0.771101, 0.566393, -0.0551891, 0.604525, 0.794672, -0.955171, 0.199883, -0.218391, 4.07625, 5.90386, -1.00545)
 light_negative = false
 light_specular = 1.0
 light_energy = 1.0
@@ -76,7 +76,7 @@ light_energy = 1.0
 [node name="Camera" type="Camera" parent="."]
 
 far = 100.0
-near = 0.10000000149011612
+near = 0.1
 projection = 0
-fov = 49.13434207760448
-transform = Transform(  0.6859206557273865, -0.32401350140571594, 0.6515582203865051, 0.0, 0.8953956365585327, 0.44527143239974976, -0.7276763319969177, -0.3054208755493164, 0.6141703724861145, 7.481131553649902, 5.34366512298584, 6.5076398849487305 )
+fov = 49.1343
+transform = Transform(0.685921, -0.324014, 0.651558, 0.0, 0.895396, 0.445271, -0.727676, -0.305421, 0.61417, 7.48113, 5.34367, 6.50764)
